zfs: Use a subslot to differ from zfsonlinux/zfs
bemerge needs libbe and grub needs beadm.
They both depend on any zfs version,
but need to depend on one with libbe and beadm.

Use a 0/libbe slot (slot 0 with subslot libbe) in zfs
to have a different version
that cannot be installed together with zfs:0/0
(which is the default that is used upstream).

Make bemerge and grub depend on zfs of this slot.