zfs-kmod: Use zfs' subslot for a matching version
zfs-kmod should be kept in sync with zfs because the kernel module's
interface might change in an incompatible way.

If zfs just depends on zfs-kmod of the same version,
upstream zfs-kmod (that uses zfsonlinux/zfs) qualifies as well.
This can be a problem when upstream code
diverges enough from linux-be/zfs to cause an incompatibility.

Make zfs-kmod use the same subslot as zfs
and make zfs depend on it
to ensure they are built from the same codebase.