#6 sys-kernel/genkernel: Use git-r3
git-2 is deprecated.

By default, git-2 uses the "master" branch;
git-r3 uses upstream HEAD.

We need the latter to use the "beadm" branch,
which is the default branch in gitlab.com/linux-be/genkernel.git.

This is better than hardcoding the "beadm" branch with EGIT_BRANCH:
a user can override the repository with genkernel_LIVE_REPO
without also having to override the branch.

Overriding ${S} is no longer necessary,
as git-r3 checks out the sources to ${WORKDIR}/${P} by default -
the same as for a release tarball.

Closes #6.