#8 sys-fs/zfs*: 0.8.4*: Fetch from a git tag
Don't inherit linux-mod in zfs-0.8.4-r1.ebuild,
because the zfs*-0.8.4* ebuilds are not live ebuilds -
they are built from the same code on every rebuild,
so it's enough to rebuild the -kmod only for a new kernel.

Set Gentoo release, since the zfs (not including libbe) code
is effectively the same as the gentoo ebuild builds.