Check linux-be invariants
Ensuring that a merge of upstream
does not overwrite the linux-be-specific changes
has turned out to be error-prone and hard to do systematically.

At least before a better process of merging upstream /
resolving conflicts is conceived of,
a script for checking if the important changes are present is useful.

Add a sanity.sh script that does that and run it in CI.