sanity.sh: Update zfs dependency check
Check that the program ebuilds mention beadm and zfs:0/libbe
instead of checking that they don't mention regular zfs.

Still, apply the old check to the zfs ebuilds,
which should depend on zfs-kmod:0/libbe
to ensure the kmod ebuild is in sync with the zfs ebuild.

Don't apply any check for sys-apps/beadm:
it should depend on a regular zfs ebuild
(without beadm and libbe included),
so this kind of check is not relevant.